Rework init
A bool was used to discriminate wether it was the first time the program was running
or was already a real check to see if kstars is alive.
Now the program checks immediately before starting the loop if we have kstars available
and the bool flag has no sense in this context.
Additionally we are preparing to use the process PID to work further and speed things up
when looking for processes.